update schema and design to be more up to date with new design
